Use smallint for runner_type since its an enum

+++ db/migrate/20180430101916_add_runner_type_to_ci_runners.rb
@@ -4,6 +4,6 @@ class AddRunnerTypeToCiRunners < ActiveRecord::Migration
   DOWNTIME = false
 
   def change
-    add_column :ci_runners, :runner_type, :integer
+    add_column :ci_runners, :runner_type, :smallint
   end
 end
